\name{mix_moment-methods}
\docType{methods}
%\alias{mix_moment}
\alias{mix_moment-methods}
\alias{mix_moment,mixAR,missing,missing,numeric-method}

\title{Methods for mix_moment}
\description{Methods for \code{mix_moment} in package \pkg{pcts}}

\section{Methods}{
\describe{

\item{\code{signature(model = "mixAR", x = "missing", index = "missing", xcond = "numeric")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link{mix_location}} for examples and related functions
}
\keyword{methods}
\keyword{ mixAR }
\keyword{internal}
