\name{mix_variance-methods}
\docType{methods}
%\alias{mix_variance}
\alias{mix_variance-methods}
\alias{mix_variance,MixAR,missing,missing,missing-method}
\alias{mix_variance,MixAR,missing,missing,numeric-method}
\alias{mix_variance,MixAR,numeric,numeric,missing-method}

\title{Methods for mix_variance}
\description{Methods for \code{mix_variance} in package \dQuote{pcts}}

\section{Methods}{
\describe{

\item{\code{signature(model = "MixAR", x = "missing", index = "missing", xcond = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(model = "MixAR", x = "missing", index = "missing", xcond = "numeric")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(model = "MixAR", x = "numeric", index = "numeric", xcond = "missing")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link{mix_location}} for examples and related functions
}
\keyword{methods}
\keyword{ MixAR }
\keyword{internal}
