\name{mixAR-internal}
\alias{param_score_stdt}
% \alias{fn_stdt}
% \alias{fdist_stdnorm}
%\alias{em_est_dist}

% \alias{}

%% Part of the sensible export list but (currently?) documented as
%% internal (or waiting for documentation to be written):
%% <LIST>
%% </LIST>

\title{Internal mixAR Functions}
\description{
  Internal mixAR functions
}
\usage{
param_score_stdt(x, nu)
}
\arguments{
  \item{x}{~~ TODO: describe this argument. ~~}
  \item{nu}{~~ TODO: describe this argument. ~~}
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written.
}
\keyword{internal}
