\name{mixARnoise_sim}
\alias{mixARnoise_sim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Simulate white noise series from a list of functions and vector of regimes
}
\description{
  Simulate white noise series from a list of functions and vector of
  regimes. This function is used internally for simulation from  mixAR models.
}
\usage{
mixARnoise_sim(rdist, z)
}
\arguments{
  \item{rdist}{a list of functions for random number generation, see `Details'.}
  \item{z}{a vector of positive integers specifying the 'regimes'.}
}
\details{
  If the length of the list \code{rdist} is \code{max(z)}, then
  \code{z[[i]]} is the random number generator for regime \code{i}.
  Alternatively, if \code{rdist} is of length one, then the same
  generator will be used for all regimes.

  \code{mixARnoise_sim} returns a vector, say \code{y}, of the same
  length as \code{z}, such that \code{y[i]} is generated by
  \code{z[[i]]}.
}
\value{
  a numeric vector
}




%% ~Make other sections like Warning with \section{Warning }{....} ~



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{simulation}
