\name{noise_dist-methods}
\docType{methods}
\alias{noise_dist-methods}
\alias{noise_dist,MixAR-method}
\alias{noise_dist,MixARGaussian-method}
\alias{noise_dist,MixARgen-method}
\title{ Methods for function \code{noise_dist} in package \pkg{mixAR} }
\description{
  Methods for function \code{noise_dist} in package \pkg{mixAR}
}
\section{Methods}{
\describe{

\item{\code{signature(model = "MixAR")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(model = "MixARGaussian")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(model = "MixARgen")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link{noise_dist}}
}
\keyword{methods}
