\name{noise_params-methods}
\docType{methods}
\alias{noise_params-methods}
\alias{noise_params,MixAR-method}
\alias{noise_params,MixARgen-method}
\title{ Methods for function \code{noise_params} in package \pkg{mixAR} }
\description{
  Methods for function \code{noise_params} in package \pkg{mixAR}
}
\section{Methods}{
\describe{

\item{\code{signature(model = "MixAR")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(model = "MixARgen")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
