\name{noise_rand-methods}
\docType{methods}
\alias{noise_rand-methods}
\alias{noise_rand,MixAR-method}
\alias{noise_rand,MixARGaussian-method}
\alias{noise_rand,MixARgen-method}
\title{ Methods for function \code{noise_rand} in package \pkg{mixAR} }
\description{
  Methods for function \code{noise_rand} in package \pkg{mixAR}
}
\section{Methods}{
\describe{

\item{\code{signature(model = "MixAR")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(model = "MixARGaussian")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(model = "MixARgen")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
