\name{tau2probhat}
\alias{tau2probhat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Estimate probabilities of a mixAR model from tau.
}
\description{
  Estimate probabilities of a mixAR model from tau.
}
\usage{
tau2probhat(tau)
}
\arguments{
  \item{tau}{todo: explain}
}

\value{
  numeric vector
}




%% ~Make other sections like Warning with \section{Warning }{....} ~



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{math}
\keyword{internal}
