\name{mix_moment-methods}
\docType{methods}
%\alias{mix_moment}
\alias{mix_moment-methods}
\alias{mix_moment,mixAR,missing,missing,numeric-method}
\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Methods}{
\describe{

\item{\code{signature(model = "mixAR", x = "missing", index = "missing", xcond = "numeric")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link{mix_location}} for examples and related functions
}
\keyword{methods}
\keyword{ mixAR }
\keyword{internal}
