\name{mix_variance-methods}
\docType{methods}
%\alias{mix_variance}
\alias{mix_variance-methods}
\alias{mix_variance,mixAR,missing,missing,missing-method}
\alias{mix_variance,mixAR,missing,missing,numeric-method}
\alias{mix_variance,mixAR,numeric,numeric,missing-method}
\title{~~ Dummy title ~~}
\description{~~ Dummy description ~~}
\section{Methods}{
\describe{

\item{\code{signature(model = "mixAR", x = "missing", index = "missing", xcond = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(model = "mixAR", x = "missing", index = "missing", xcond = "numeric")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(model = "mixAR", x = "numeric", index = "numeric", xcond = "missing")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link{mix_location}} for examples and related functions
}
\keyword{methods}
\keyword{ mixAR }
\keyword{internal}
