\name{noise_dist-methods}
\docType{methods}
\alias{noise_dist-methods}
\alias{noise_dist,mixAR-method}
\alias{noise_dist,mixARGaussian-method}
\alias{noise_dist,mixARgen-method}
\title{ Methods for function \code{noise_dist} in package \pkg{mixAR} }
\description{
  Methods for function \code{noise_dist} in package \pkg{mixAR}
}
\section{Methods}{
\describe{

\item{\code{signature(model = "mixAR")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(model = "mixARGaussian")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(model = "mixARgen")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link{noise_dist}}
}
\keyword{methods}
